\name{SIRTGP-package}
\alias{SIRTGP-package}
\alias{SIRTGP}
\docType{package}
\title{
\packageTitle{SIRTGP}
}
\description{
\packageDescription{SIRTGP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SIRTGP}
\packageIndices{SIRTGP}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SIRTGP}

Maintainer: \packageMaintainer{SIRTGP}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
